Moved some tests to different classes.
