+++ src/test/java/io/jenkins/plugins/analysis/warnings/StepsITest.java
@@ -18,6 +18,7 @@ import edu.hm.hafner.analysis.Issue;
 import edu.hm.hafner.analysis.Report;
 
 import org.kohsuke.stapler.HttpResponse;
+import org.jenkinsci.Symbol;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
@@ -28,6 +29,7 @@ import hudson.util.HttpResponses;
 import io.jenkins.plugins.analysis.core.model.AnalysisResult;
 import io.jenkins.plugins.analysis.core.model.ReportScanningTool;
 import io.jenkins.plugins.analysis.core.model.ResultAction;
+import io.jenkins.plugins.analysis.core.model.Tool;
 import io.jenkins.plugins.analysis.core.steps.PublishIssuesStep;
 import io.jenkins.plugins.analysis.core.steps.ScanForIssuesStep;
 import io.jenkins.plugins.analysis.core.testutil.IntegrationTestWithJenkinsPerTest;
@@ -46,6 +48,14 @@ import static io.jenkins.plugins.analysis.core.assertions.Assertions.*;
  */
 @SuppressWarnings("PMD.ExcessiveImports")
 public class StepsITest extends IntegrationTestWithJenkinsPerTest {
+    /** Verifies that a {@link Tool} defines a {@link Symbol}. */
+    @Test
+    public void shouldProvideSymbol() {
+        FindBugs findBugs = new FindBugs();
+
+        assertThat(findBugs.getSymbolName()).isEqualTo("findBugs");
+    }
+
     /**
      * Creates a declarative Pipeline and scans for a Gcc warning.
      */
